Rename faction police to polizei in packages
